+++ test/controllers/messages_controller_test.rb
@@ -99,17 +99,6 @@ class MessagesControllerTest < ActionDispatch::IntegrationTest
     assert_equal @user.contactable, false
   end
 
-  test "#incoming should handle incoming message with pause" do
-    create(:auto_response, trigger_phrase: "pause", response: "Thanks, you've paused for 4 weeks.", update_user: '{"contactable": false, "restart_at": "4.weeks.from_now.noon"}')
-    MessagesController.any_instance.stubs(:valid_twilio_request?).returns(true)
-
-    post messages_incoming_url, params: {From: @user.phone_number, Body: "Pause ", MessageSid: "new_sid"}
-    assert_response :success
-    @user.reload
-    refute @user.contactable
-    refute_nil @user.restart_at
-  end
-
   test "#incoming should handle incoming message with start" do
     create(:auto_response, trigger_phrase: "start", update_user: "{\"contactable\": true}")
     @user.update(contactable: false)