+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [2026-04-08]
+- Offer to put users on a waitlist if their child isn't old enough
+- Send them a text to register once their child is 9 months old
+
 ## [2026-04-07]
 - Put a link to pre-programme survey on the thank you page
 - Send reminders to users to fill it in if they haven't a couple of days later